# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Database API
  version: 1.0.0
  description: |
    # Knowledge Database API
    
    Comprehensive knowledge management system with AI-powered search and bilingual support.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (viewer, editor, admin)
    - Multi-language support (English and Korean)
    - AI-powered semantic search using OpenSearch
    - Real-time analytics and metrics
    - Rate limiting and request throttling
    - Comprehensive audit logging
    
    ## Authentication
    Most endpoints require authentication using JWT bearer tokens.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - Default: 100 requests per 60 seconds per user
    - Burst: Maximum 10 requests per second
    
    ## Language Support
    Use the `language` query parameter to specify language preference:
    - `en` for English (default)
    - `ko` for Korean
  contact:
    email: api-support@knowledge.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.knowledge.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
...
paths:
...
  /search:
    post:
      summary: Search knowledge base
      tags: [Search]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LanguageParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search_type
          in: query
          schema:
            type: string
            enum: [keyword, semantic, hybrid]
            default: hybrid
          description: Search type
        - name: category_ids
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by category IDs
        - name: types
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by content types
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Search query
                  example: how to configure authentication
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /search/suggest:
    get:
      summary: Get search suggestions
      tags: [Search]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LanguageParam'
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Partial search query
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Maximum suggestions
      responses:
        '200':
          description: Search suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestResponse'
  /search/similar:
    post:
      summary: Find similar content
      tags: [Search]
      parameters:
        - $ref: '#/components/parameters/LanguageParam'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Maximum results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to find similar content for
                  example: I need help with user authentication and JWT tokens
      responses:
        '200':
          description: Similar items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    LanguageParam:
      name: language
      in: query
      schema:
        type: string
        enum: [en, ko]
        default: en
      description: Language preference
  schemas:
...
    CategoryInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name_ko:
          type: string
        name_en:
          type: string
        slug:
          type: string
...
    SearchResponse:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              title:
                type: string
              slug:
                type: string
              summary:
                type: string
              score:
                type: number
                format: float
              highlights:
                type: array
                items:
                  type: string
              category:
                $ref: '#/components/schemas/CategoryInfo'
              tags:
                type: array
                items:
                  type: string
        facets:
          type: object
          properties:
            categories:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  count:
                    type: integer
            tags:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
    SuggestResponse:
      type: object
      properties:
        query:
          type: string
        suggestions:
          type: array
          items:
            type: string
    SimilarResponse:
      type: object
      properties:
        text:
          type: string
        language:
          type: string
        similar_items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              similarity:
                type: number
                format: float
...
  responses:
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              retry_after:
                type: integer
          example:
            error: rate_limit_exceeded
            message: Rate limit exceeded. Please try again later.
            retry_after: 45
...